RF: Moved around some code to make it more workable for future additions.
