+++ psychopy/app/coder/styling.py
